requirements.txt
fastapi==0.75.0
pika==1.1.0
Uvicorn[standard]
python-dotenv
requests
sendgrid